refactor: Reformat code, optimize imports.
