+++ src/classes/commands/SubCommand.ts
@@ -1,7 +1,7 @@
 // Note : The SubCommand class is in the Command.ts file, see the class to know why.
 
-import {Tag} from './Command.js';
-import {SubCommandContext} from '../contexts/SubCommandContext.js';
+import {SubCommandContext} from '../contexts';
+import {Tag} from './Command';
 
 export type RunSubCommandFunction = (ctx: SubCommandContext) => Promise<any>;
