+++ src/classes/contexts/EventContext.ts
@@ -1,5 +1,5 @@
-import {CommandHandler} from '../../CommandHandler.js';
-import {Event} from '../Event.js';
+import {CommandHandler} from '../../CommandHandler';
+import {Event} from '../Event';
 
 interface EventContextBuilder<E extends Event> {
 	event: E;
@@ -15,11 +15,11 @@ export class EventContext<E extends Event> implements EventContextBuilder<E> {
 		this.handler = options.handler;
 	}
 
-	get eventName() {
-		return this.event.name;
-	}
-
 	get client() {
 		return this.handler.client!!;
+	}
+
+	get eventName() {
+		return this.event.name;
 	}
 }
