+++ src/utils/JSONUtils.ts
@@ -1,5 +1,5 @@
 import * as fs from 'fs';
-import {Logger} from './Logger.js';
+import {Logger} from './Logger';
 
 export type JSONLike = {[k: string]: number | string | boolean | JSONLike | JSONLike[]};
