+++ src/utils/getThing.ts
@@ -1,5 +1,5 @@
 import {Channel, Collection, Emoji, Guild, GuildChannel, GuildMember, Message, NewsChannel, Role, Snowflake, TextChannel, User} from 'discord.js';
-import {Command} from '../classes/commands/Command.js';
+import {Command} from '../classes';
 import {CommandHandler} from '../CommandHandler';
 
 export enum DataType {
